Adds markdown for some fields
