+++ app/styles/index.scss
@@ -2,6 +2,7 @@
   "mixins",
   "settings",
   "base",
+  "markdown",
 
   "layouts",
   "components";
