+++ package.json
@@ -73,6 +73,7 @@
     "react-dom": "15.6.2",
     "react-geosuggest": "2.6.0",
     "react-leaflet": "1.7.8",
+    "react-markdown": "^3.1.1",
     "react-redux": "4.4.8",
     "react-redux-toastr": "^7.1.6",
     "react-router": "2.8.1",
